# IndianMasterApp-Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        description: Pagination metadata or other metadata
      success:
        type: boolean
    type: object
...
  dto.ApproveVerificationRequest:
    properties:
      entityId:
        type: string
      entityType:
        enum:
        - WORKER
        - HIRER
        - BUSINESS
        type: string
    required:
    - entityId
    - entityType
    type: object
...
  dto.CreateApplicationRequest:
    properties:
      jobId:
        type: string
    required:
    - jobId
    type: object
...
  dto.CreateJobRequest:
    properties:
      addressText:
        maxLength: 1000
        type: string
      benefits:
        items:
          type: string
        type: array
      businessId:
        type: string
      categories:
        items:
          type: string
        type: array
      city:
        maxLength: 100
        type: string
      experienceMax:
        minimum: 0
        type: integer
      experienceMin:
        minimum: 0
        type: integer
      jobRole:
        maxLength: 100
        type: string
      position:
        maxLength: 200
        type: string
      preferredLanguages:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
      salaryMaxAmount:
        minimum: 0
        type: number
      salaryMinAmount:
        minimum: 0
        type: number
      state:
        maxLength: 100
        type: string
      vacancies:
        minimum: 1
        type: integer
      weeklyLeaves:
        maximum: 7
        minimum: 0
        type: integer
      workType:
        maxLength: 50
        type: string
      workingHours:
        maximum: 24
        minimum: 0
        type: integer
    required:
    - businessId
    - city
    - jobRole
    - state
    - vacancies
    type: object
...
  dto.RejectVerificationRequest:
    properties:
      entityId:
        type: string
      entityType:
        enum:
        - WORKER
        - HIRER
        - BUSINESS
        type: string
      reason:
        maxLength: 1000
        type: string
    required:
    - entityId
    - entityType
    - reason
    type: object
...
  dto.UpdateApplicationStatusRequest:
    properties:
      status:
        enum:
        - shortlisted
        - rejected
        - accepted
        - withdrawn
        type: string
    required:
    - status
    type: object
...
info:
  contact:
    email: support@indianmaster.com
    name: API Support
  description: Backend APIs for Indian Master Application - A comprehensive platform
    for job seekers and hirers in India
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Indian Master API
  version: "1.0"
paths:
  /admin/applications/{application_id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a job application. Notifies the worker of
        the status change.
      parameters:
      - description: Application ID
        in: path
        name: application_id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateApplicationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Application status updated successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to update application status
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Application Status
      tags:
      - Application
...
  /admin/verification/approve:
    post:
      consumes:
      - application/json
      description: Approve a pending verification request. Notifies the worker/business
        owner.
      parameters:
      - description: Verification approval details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApproveVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification approved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden - Admin role required
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to approve verification
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve Verification
      tags:
      - Admin
  /admin/verification/reject:
    post:
      consumes:
      - application/json
      description: Reject a pending verification request. Notifies the worker/business
        owner with the reason.
      parameters:
      - description: Verification rejection details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RejectVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification rejected successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden - Admin role required
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to reject verification
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Reject Verification
      tags:
      - Admin
...
  /applications:
    post:
      consumes:
      - application/json
      description: Submit a job application. Notifies the hirer and invalidates recommendation
        cache.
      parameters:
      - description: Application details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Application submitted successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "409":
          description: Already applied to this job
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to apply for job
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Apply to Job
      tags:
      - Application
  /applications/my-applications:
    get:
      consumes:
      - application/json
      description: Retrieve all applications submitted by the authenticated worker
        with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Applications retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch applications
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get My Applications
      tags:
      - Application
...
  /hirer/jobs:
    post:
      consumes:
      - application/json
      description: Create a new job posting. Invalidates jobs feed cache and notifies
        matching workers.
      parameters:
      - description: Job details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Job created successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to create job
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Job Posting
      tags:
      - Job
  /hirer/jobs/{job_id}:
    put:
      consumes:
      - application/json
      description: Update an existing job posting. Invalidates jobs feed, search,
        and recommendation caches.
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      - description: Updated job details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Job updated successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to update job
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Job Posting
      tags:
      - Job
  /jobs/{job_id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific job
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Job Details
      tags:
      - Job
  /jobs/feed:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated feed of available jobs. Results are cached
        in Redis for 5 minutes.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by job role
        in: query
        name: job_role
        type: string
      - description: Filter by work type
        in: query
        name: work_type
        type: string
      - description: Minimum salary
        in: query
        name: salary_min
        type: integer
      - description: Maximum salary
        in: query
        name: salary_max
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Jobs feed retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch jobs feed
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Jobs Feed (Cached)
      tags:
      - Job
...
  /worker/profile:
    get:
      consumes:
      - application/json
      description: Retrieve the worker profile for the authenticated user (Worker
        Only)
      produces:
      - application/json
      responses:
        "200":
          description: Worker profile retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Worker profile not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Worker Profile
      tags:
      - Worker
    post:
      consumes:
      - application/json
      description: Create a new worker profile for the authenticated user (Worker
        Only)
      parameters:
      - description: Worker profile details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWorkerProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Worker profile created successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to create worker profile
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Worker Profile
      tags:
      - Worker
    put:
      consumes:
      - application/json
      description: Update an existing worker profile. Invalidates Redis cache for
        recommendations and search.
      parameters:
      - description: Updated worker profile details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWorkerProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Worker profile updated successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to update worker profile
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Worker Profile
      tags:
      - Worker
  /worker/profile/verification:
    get:
      consumes:
      - application/json
      description: Retrieve the verification status for a worker (Worker Only)
      parameters:
      - description: Worker ID
        in: path
        name: worker_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Verification status retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Verification status not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Verification Status
      tags:
      - Worker
...
schemes:
- http
- https
swagger: "2.0"